rawger/architecture: Localize DefinerView, change selectedTab dataflow
Major refactor to have DefinerView wrap only the views necessary, as opposed to all of app central. This allows navigation stacks to have different definitions in each frame. Also changes the selectedTab dataflow to hopefully minimize rerendering.

These changes inadvertently fix the issue where there would be an empty space above the bottomsheet on the first render of the first view. Also inadvertently introduces a nice transition animation when the definition changes.
